requirements.txt
torch
torchvision
numpy
Pillow
tqdm
albumentations
matplotlib
seaborn
scikit-learn
